+++ zappy_server/src/commands/commands_ia/funct_client_ia_eject.c
@@ -13,7 +13,7 @@ void funct_client_ia_eject(ia_t *ia, uint8_t **args, common_t *com)
 
     (void)com;
     (void)args;
-    new_msg->time = 7 / com->freq;
+    new_msg->time = 7. / (double)com->freq;
     new_msg->start = 0;
     new_msg->msg = NULL;
     new_msg->handler = &funct_response_ia_eject;
